try to debug fetchMock in sunsetDataTest
* experimented with other fetchMock methods:
- When I chained a .catch() onto the mock call, 
I was able to get through to some of the assertions in the test. 
Before this, I got an error that  
"No fallback response defined for GET” to the url I specified
- when I chain a .spy on the mock call (which records 
calls made to the url specified but sends request to fetch() ), 
I get a "Maximum call stack size exceeded” error. 
this happens even when I specify a number of times 
to call the api in the options object
* logging the response of the fetch call in the 
actual function leads to something  weird: the 
response returns a 200 status, but the details of the 
body have nothing to do with sunrise data…not sure
what is going on 
* next steps are to try hitting this api with another
http request library